Move CassandraUtils to src; CassandraConfigurator: move static read-only props not to create mess in the beginning of the class
